+++ src/org/sosy_lab/cpachecker/util/floatingpoint/test/CFloatUnitTest.java
@@ -214,7 +214,7 @@ public abstract class CFloatUnitTest {
   }
 
   private static int calculateExpWidth(BinaryMathContext pFormat) {
-    return (int) Math.ceil(lb((double) (2 * pFormat.maxExponent + 1)));
+    return (int) Math.ceil(lb(2 * pFormat.maxExponent + 1));
   }
 
   private static String toBits(BinaryMathContext format, BigFloat value) {
